Enable deprecation warnings and disable verbose mode
Deprecation warnings are easily lost so make them a bit more obvious
in the test output.

Verbose mode is not needed for Jenkins which will be using JUnit XML.
It can be added by users by having a setup.cfg. If we want to allow
use of $PYTEST_ADDOPTS we may have to pass that through the environment.