Be explicit about which python pytest-xdist uses.
When starting a Python interpreter, it will add its own site-packages
directory to sys.path if it's not already there.

Normally, this works fine. However, when the same interpreter can be accessed
through two different paths, it can cause both of those paths to be added to
sys.path.

For example, in our standard, newinstall.sh provided, environment, we might
refer to the Python interpreter either as:

  /ssd/lsstsw/stack3_20171021/python/miniconda3-4.3.21/bin/python

or as:

  /ssd/lsstsw/stack3_20171021/python/current/bin/python

If we include the former's site packages directory corresponding to the former
on our $PYTHONPATH and then run it, everything is fine. Howver, if we instead
invoke the latter, it will insert a new site-packages directory,
corresponding to that provided by Miniconda, to our sys.path. It doesn't add
it at the end of the list of paths, but rather inserts it at some (arbitrary?)
point in the middle.

This then causes confusion when there are packages installed by Miniconda
which shadow packages provided by EUPS: even if the latter are set up, they
can be overriden by this new entry on sys.path. Confusion ensues.

Here, we explicitly normalize the path used by pytest-xdist to invoke Python
so that we avoid this problem.