Add very limited compile_commands.json support.
This only writes out the compile commands that are actually being run
(i.e. not files that are intentionally not being rebuilt) and only when
'compile_commands.json' is listed explicitly as a target.

That makes this *barely* usable for editor integration; if you start
from a clean repo, you can build from scratch with that target
included, making your editor happy, and then rebuild as usual (without
that target) to avoid clobbering the file with a version that includes
much less.

Unfortunately, doing better than this would seem to involve
understanding much more of SCons' internals than I have time for now,
and I worry that it's not even a well-formed operation, given the
degree to which sconsUtils regularly disregards the super-important
distinction between scons-execution-time and build-time logic.  More
time is probably better invested in just switching us to CMake.