Fix handling of -I, -isystem, and -L from envvars.
Previously -isystem was ignored in CFLAGS and CXXFLAGS, while only the
-I and -L options were stripped, not their arguments.  It seems we
regularly only get -isystem options in these envvars from conda, which
was why this was not noticed.

It's also important to add these paths back in to the appropriate
construction variables, especially now that this logic is actually
doing something.

By putting these paths in these construction variables, we make it
possible to put new paths in front of them, which is the only way we
could ever override something conda provides (which is a somewhat scary
thing to do for ABI reasons, but it was important in the past for Eigen
and is important for pybind11 now, and as header-only libraries those
are easier to handle in terms of ABI).